refactor sample data code
