+++ src/App.js
@@ -10,7 +10,6 @@ import {
 } from "little-state-machine";
 
 createStore({
-  yourDetail: { name: "" },
   formSchemaState: [],
   formContainerState: [],
   formValuesState: [],
