+++ src/Components/sampleData/sampleFormValues.js
@@ -1,0 +1,6 @@
+export let sampleInitialFormValues = [
+  { key: "abc1", fieldValue: "" },
+  { key: "abc2", fieldValue: "" },
+  { key: "abc3", fieldValue: "" },
+  { key: "abc4", fieldValue: "" },
+];
